# back_end/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Swagger Fotos - OpenAPI 3.0
  version: 1.0.11
servers:
  - url: http://localhost:8000
  - url: http://localhost:8001
tags:
  - name: user
    description: Operations about user
paths:
  /api/auth/sign-in:
    post:
      tags:
        - user
      summary: Logs user into the system
      description: ''
      operationId: loginUser
      security:
        - api_key: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInBody'
      responses:
        '200':
          description: successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
             application/json:
              schema:
                $ref: '#/components/schemas/SuccessAuthentication'
        '400':
          description: Invalid username/password supplied
  /api/auth/sign-up:
    post:
      tags:
        - user
      summary: Register new account
      description: ''
      operationId: registerUser
      security:
        - api_key: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpBody'
      responses:
        '200':
          description: successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
             application/json:
              schema:
                $ref: '#/components/schemas/SuccessAuthentication'
        '400':
          description: Invalid username/password supplied
  /api/user/upload-avatar:
    post:
      tags:
        - user
      summary: Upload user avatar
      description: ''
      operationId: uploadAvatar
      security:
        - api_key: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadAvatarBody'
      responses:
        '200':
          description: successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
             application/json:
              schema:
                $ref: '#/components/schemas/SuccessAuthentication'
        '400':
          description: Invalid username/password supplied
components:
  schemas:
    SignInBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      xml:
        name: tag
    SignUpBody:
      type: object
      properties:
        email:
          type: string
          example: example@email.com
        username:
          type: string
          example: username
        password:
          type: string
          example: password
        confirm_password:
          type: string
          example: password
      xml:
        name: tag
    UploadAvatarBody:
      type: object
      properties:
        file:
          type: string
          format: binary
      
      xml:
        name: tag
    SuccessAuthentication:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: object
          properties:
            access_token:
              type: string
              example: 'eyJhbGciOi'
            refresh_token:
              type: string
              example: 'eyJhbGciOi'
            exprires_access_token:
              type: string
              example: '1d'
      xml:
        name: user
  securitySchemes:
    api_key:
      type: apiKey
      name: x-api-key
      in: header